+++ tests/test_api/test_subscriptions.py
@@ -447,3 +447,118 @@ def test_post_subscriptions_validate_only(client, tables):
 
     assert response.status_code == HTTPStatus.OK
     assert len(tables.subscriptions_table.scan()['Items']) == 1
+
+
+def test_query_subscription_by_name(client, tables):
+    login(client, 'subscriptionsUser')
+    items = [
+        {
+            'subscription_id': 'f00b731f-121d-44dc-abfa-c24afd8ad542',
+            'user_id': 'subscriptionsUser',
+            'creation_date': '2020-01-01T00:00:00+00:00',
+            'search_parameters': {
+                'start': '2020-01-01T00:00:00+00:00',
+                'end': '2020-01-01T00:00:00+00:00',
+            },
+            'job_specification': {
+                'job_type': 'RTC_GAMMA',
+                'name': 'SubscriptionName1'
+            }
+        },
+        {
+            'subscription_id': '140191ab-486b-4080-ab1b-3e2c40aab6b8',
+            'user_id': 'subscriptionsUser',
+            'search_parameters': {
+                'start': '2020-01-01T00:00:00+00:00',
+                'end': '2020-01-01T00:00:00+00:00',
+            },
+            'job_specification': {
+                'job_type': 'RTC_GAMMA',
+                'name': 'SubscriptionName2'
+            }
+        }
+    ]
+
+    for item in items:
+        tables.subscriptions_table.put_item(Item=item)
+
+    response = client.get(SUBSCRIPTIONS_URI, query_string={'name': 'SubscriptionName1'})
+    assert response.status_code == HTTPStatus.OK
+    assert response.json['subscriptions'] == [items[0]]
+
+
+def test_query_jobs_by_job_type(client, tables):
+    login(client, 'subscriptionsUser')
+    items = [
+        {
+            'subscription_id': 'f00b731f-121d-44dc-abfa-c24afd8ad542',
+            'user_id': 'subscriptionsUser',
+            'creation_date': '2020-01-01T00:00:00+00:00',
+            'search_parameters': {
+                'start': '2020-01-01T00:00:00+00:00',
+                'end': '2020-01-01T00:00:00+00:00',
+            },
+            'job_specification': {
+                'job_type': 'RTC_GAMMA',
+                'name': 'SubscriptionName1'
+            }
+        },
+        {
+            'subscription_id': '140191ab-486b-4080-ab1b-3e2c40aab6b8',
+            'user_id': 'subscriptionsUser',
+            'search_parameters': {
+                'start': '2020-01-01T00:00:00+00:00',
+                'end': '2020-01-01T00:00:00+00:00',
+            },
+            'job_specification': {
+                'job_type': 'INSAR_GAMMA',
+                'name': 'SubscriptionName1'
+            }
+        }
+    ]
+
+    for item in items:
+        tables.subscriptions_table.put_item(Item=item)
+
+    response = client.get(SUBSCRIPTIONS_URI, query_string={'job_type': 'RTC_GAMMA'})
+    assert response.status_code == HTTPStatus.OK
+    assert response.json['subscriptions'] == [items[0]]
+
+def test_query_jobs_by_enabled(client, tables):
+    login(client, 'subscriptionsUser')
+    items = [
+        {
+            'subscription_id': 'f00b731f-121d-44dc-abfa-c24afd8ad542',
+            'user_id': 'subscriptionsUser',
+            'creation_date': '2020-01-01T00:00:00+00:00',
+            'enabled': True,
+            'search_parameters': {
+                'start': '2020-01-01T00:00:00+00:00',
+                'end': '2020-01-01T00:00:00+00:00',
+            },
+            'job_specification': {
+                'job_type': 'RTC_GAMMA',
+                'name': 'SubscriptionName1'
+            }
+        },
+        {
+            'subscription_id': '140191ab-486b-4080-ab1b-3e2c40aab6b8',
+            'user_id': 'subscriptionsUser',
+            'enabled': False,
+            'search_parameters': {
+                'start': '2020-01-01T00:00:00+00:00',
+                'end': '2020-01-01T00:00:00+00:00',
+            },
+            'job_specification': {
+                'job_type': 'INSAR_GAMMA',
+                'name': 'SubscriptionName1'
+            }
+        }
+    ]
+
+    for item in items:
+        tables.subscriptions_table.put_item(Item=item)
+
+    response = client.get(SUBSCRIPTIONS_URI, query_string={'enabled': True})
+    assert response.status_code == HTTPStatus.OK
+    assert response.json['subscriptions'] == [items[0]]